No quote wrapping for non-sysctl sections
Although quoting space-containing values in the `[sysctl]` section of a
tuned profile is fine, arguably necessary, and shown in RedHat's
included default tuning profiles, when our template adds quotes to all
values in all sections, they are taken literally and included as if they
were part of the value itself.  Adjust this quote-by-default logic in
the profile template so that it only quotes values inside of the
`[sysctl]` section, and nothing else.  It is still possible for anyone
who needs literal quotes around a value outside of this section to pass
them as part of Puppet's value, but the template will no longer add
them.